backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
numpy
opencv-python-headless
jinja2
pytest
